tools/asset_pipeline/requirements.txt
Pillow>=10.0.0
PyYAML>=6.0
numpy>=1.24.0
rembg>=2.0.50
